+++ javascript/packages/eslint-plugin-chotot/rules/no-abusive-eslint-disable.js
@@ -1,13 +1,15 @@
-/**
- * Copyright (c) 2016 Sindre Sorhus <sindresorhus@gmail.com> (https://sindresorhus.com)
- */
 'use strict';
-const getDocumentationUrl = require('./utils/get-documentation-url');
 
-const disableRegex = /^eslint-disable(?:-next-line|-line)?(?<ruleId>$|(?:\s+(?:@(?:[\w-]+\/){1,2})?[\w-]+)?)/;
+const MESSAGE_ID = 'no-abusive-eslint-disable';
+const messages = {
+  [MESSAGE_ID]: 'Specify the rules you want to disable.',
+};
 
-const create = (context) => ({
-  Program: (node) => {
+const disableRegex =
+  /^eslint-disable(?:-next-line|-line)?(?<ruleId>$|(?:\s+(?:@(?:[\w-]+\/){1,2})?[\w-]+)?)/;
+
+const create = () => ({
+  *Program(node) {
     for (const comment of node.comments) {
       const value = comment.value.trim();
       const result = disableRegex.exec(value);
@@ -16,7 +18,9 @@ const create = (context) => ({
         result && // It's a eslint-disable comment
         !result.groups.ruleId // But it did not specify any rules
       ) {
-        context.report({
+        yield {
+          // Can't set it at the given location as the warning
+          // will be ignored due to the disable comment
           loc: {
             start: {
               ...comment.loc.start,
@@ -24,8 +28,8 @@ const create = (context) => ({
             },
             end: comment.loc.end,
           },
-          message: 'Specify the rules you want to disable.',
-        });
+          messageId: MESSAGE_ID,
+        };
       }
     }
   },
@@ -36,7 +40,8 @@ module.exports = {
   meta: {
     type: 'suggestion',
     docs: {
-      url: getDocumentationUrl(__filename),
+      description: 'Enforce specifying rules to disable in `eslint-disable` comments.',
     },
+    messages,
   },
 };