+++ javascript/packages/eslint-plugin-chotot/rules/selectors/empty-object-selector.js
@@ -1,0 +1,8 @@
+'use strict';
+
+function emptyObjectSelector(path) {
+  const prefix = path ? `${path}.` : '';
+  return [`[${prefix}type="ObjectExpression"]`, `[${prefix}properties.length=0]`].join('');
+}
+
+module.exports = emptyObjectSelector;