+++ javascript/packages/eslint-plugin-chotot/rules/utils/escape-template-element-raw.js
@@ -1,0 +1,4 @@
+'use strict';
+
+module.exports = (string) =>
+  string.replace(/(?<=(?:^|[^\\])(?:\\\\)*)(?<symbol>(?:`|\$(?={)))/g, '\\$<symbol>');