+++ javascript/packages/eslint-plugin-chotot/rules/utils/get-variable-identifiers.js
@@ -1,0 +1,6 @@
+'use strict';
+const { uniq } = require('lodash');
+
+// Get identifiers of given variable
+module.exports = ({ identifiers, references }) =>
+  uniq([...identifiers, ...references.map(({ identifier }) => identifier)]);