+++ javascript/packages/eslint-plugin-chotot/changelog.md
@@ -1,6 +1,15 @@
-1.0.0 / 2020-06-05
-===================
+# 2.0.0 / 2021-08-05
+
+- Add 5 new rules:
+  - `chotot/no-instanceof-array`
+  - `chotot/no-nested-ternary`
+  - `chotot/no-this-assignment`
+  - `chotot/no-unsafe-regex`
+  - `chotot/no-unused-properties`
+- Update unit testing framework from latest `eslint-plugin-unicorn`
+
+# 1.0.0 / 2020-06-05
 
 - First release with 2 rules:
-  + `chotot/filename-case`
-  + `chotot/no-abusive-eslint-disable`
+  - `chotot/filename-case`
+  - `chotot/no-abusive-eslint-disable`